requirements.txt
Django==1.6.6
Pillow==2.5.3
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-bootstrapped==2.0.0
django-appconf==0.6
django-imagekit==3.2.2
django-storages==1.1.8
django-tinymce==1.5.2
django-toolbelt==0.0.1
gunicorn==19.1.1
pilkit==1.1.9
psycopg2==2.5.3
six==1.7.3
static3==0.5.1
